VLab Issue #15928 Dynamic rotation and scaling of shapes
Added to the GUI support for the user dynamically rotating and scaling
ellipses, lines, and polygons. When any of these types of shapes are
selected, and they are marked as allowing rotation and/or rescaling (for
hazard events, if they are editable; for visual features, explicit flags
exist offering a finer-grained approach), a bounding box (rotated if the
shape was previously rotated) appears around the selected shape, and if
the mouse cursor hovers over the shape, the four bounding box corner
points and, if rescaling is allowed, four box midpoints, appear. These
act act as "manipulation points"; the user may drag them to rotate or
rescale the shape.

Refactored vertex and whole-shape movement out of the largest, most
monolithic input handler, creating delegate input handlers to deal with
these cases, similar to the new resizing and rotating input handlers.

Refined (in terms of bug removal and behavioral consistency) the
behavior of shapes when selected versus unselected. Unselected shapes
that are modifiable in some way (movable, vertex-editable, resizable,
and/or rotatable) "light up" subtly when the mouse cursor passes over
them. Selected ones also get highlighted, but more obviously, in the
same situation, and the manipulation points appear. This gives the user
immediate visual feedback as to which shapes are modifiable in some way
and how.

Changed unselected shapes to not be editable until they are selected,
i.e. one cannot edit the vertices of an unselected shape; it must be
selected first. This makes H.S. behave in a manner much more consistent
with the shape-manipulation UIs of other software.

Fixed code under the hood to handle polygonal geometries with holes in
them correctly, preparing for a changeset that will eventually be needed
to add better shape-editing tools (combining shapes, differencing,
etc.).

+++ viz/com.raytheon.uf.viz.hazards.sessionmanager/src/com/raytheon/uf/viz/hazards/sessionmanager/config/ISessionConfigurationManager.java
@@ -91,6 +91,7 @@ import com.raytheon.uf.viz.hazards.sessionmanager.originator.IOriginator;
  *                                      the configuration manager, but the work of
  *                                      tracking data layer changes is done by the
  *                                      app builder where it belongs.
+ * Sep 27, 2016 15928      Chris.Golden Changed line thickness for hazard events.
  * </pre>
  * 
  * @author bsteffen
@@ -245,7 +246,7 @@ public interface ISessionConfigurationManager<S extends ISettings> {
      * @param selected
      * @return
      */
-    public int getBorderWidth(IHazardEvent event, boolean selected);
+    public double getBorderWidth(IHazardEvent event, boolean selected);
 
     /**
      * Get the border style to use when displaying an event.
